+++ app/Http/Controllers/ProductTypeController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Services\HelpFunctions;
 use Illuminate\Http\Request;
 use Validator;
 use App\Models\ProductType;
@@ -191,7 +192,11 @@ class ProductTypeController extends Controller
 
 		$productType = ProductType::find($id);
 		if (!$productType) return response()->json(['status' => 'error', 'reason' => trans('main.error.тип-продукта-не-найден')]);
-
+		
+		if (HelpFunctions::isDemo($productType->created_at)) {
+			return response()->json(['status' => 'error', 'reason' => 'Demo data cannot be updated']);
+		}
+		
 		$rules = [
 			'name' => ['required', 'max:255', 'unique:product_types,name,' . $id],
 			'alias' => ['required', 'min:3', 'max:50', 'unique:product_types,alias,' . $id],
@@ -239,6 +244,10 @@ class ProductTypeController extends Controller
 
 		$productType = ProductType::find($id);
 		if (!$productType) return response()->json(['status' => 'error', 'reason' => trans('main.error.тип-продукта-не-найден')]);
+		
+		if (HelpFunctions::isDemo($productType->created_at)) {
+			return response()->json(['status' => 'error', 'reason' => 'Demo data cannot be deleted']);
+		}
 		
 		if (!$productType->delete()) {
 			return response()->json(['status' => 'error', 'reason' => trans('main.error.повторите-позже')]);
